requirements.txt
streamlit              # for Front End
google-generativeai    # for Generative AI Model
python-dotenv          # for Local Environment

# pip install -r requirements.txt      # Installing all the required libraries

